conditional build for each os/arch
