+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
     "rust-analyzer.linkedProjects": [
-        "./Cargo.toml"
+        "./Cargo.toml",
     ]
 }
